requirements.txt
cached-property==1.5.2
certifi==2021.10.8
cloudpickle==1.6.0
cycler==0.11.0
Cython==0.29.26
decorator==5.1.1
fonttools==4.29.0
glfw==2.5.0
GPy==1.10.0
GPyOpt @ git+git://github.com/yunshengtian/GPyOpt@5fc1188ffdefea9a3bc7964a9414d4922603e904
gym==0.19.0
h5py==3.6.0
imageio==2.14.1
importlib-metadata==4.10.1
kiwisolver==1.3.2
matplotlib==3.5.1
neat-python @ git+git://github.com/yunshengtian/neat-python@2762ab630838520ca6c03a866e8a158f592b0370
numpy==1.21.5
packaging==21.3
pandas==1.3.5
paramz==0.9.5
Pillow==9.0.0
pybind11==2.9.0
pygifsicle==1.0.5
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.5
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
scipy==1.7.3
six==1.16.0
stable-baselines3==1.4.0
torch==1.10.1
ttkbootstrap==1.5.1
typing==3.7.4.3
typing_extensions==4.0.1
zipp==3.7.0
